feat(time_utils): use relative precision wording when precision time is zero
When using precision strings, the zero-value needs to make sense for the
specified precision. If the precision is days, a zero-value of 'now'
doesn't make sense, since 'now' only makes sense in the context of zero
seconds. If there are zero days, then we are within 'today'. If there
are zero months, then we are within 'this month'.