refactor: use namespace separators for views
I went back and forth on this a lot since it's not idiomatic, but I
think the increase in clarity is worth it here. We're going to have a
lot of name collisions by the nature of the menu items sharing a lot of
similar view states, so making sure they are explicitly namespaced will
reduce mental overhead when trying to understand the intent.